added AbstractMongoDatabaseWrapperTest abstact class
