+++ com.examples.school/src/test/java/com/examples/school/mongo/MongoDatabaseWrapperTest.java
@@ -1,62 +1,15 @@
 package com.examples.school.mongo;
 
-import static org.junit.Assert.*;
 
-import java.net.UnknownHostException;
-
-import org.junit.Before;
-import org.junit.Test;
-
-import com.examples.school.Student;
-import com.examples.school.mongo.helper.MongoTestHelper;
+import com.examples.school.mongo.common.AbstractMongoDatabaseWrapperTest;
 import com.github.fakemongo.Fongo;
 import com.mongodb.MongoClient;
 
-public class MongoDatabaseWrapperTest {
-	private MongoDatabaseWrapper mongoDatabase;
-	private MongoTestHelper mongoTestHelper;
-
-	@Before
-	public void initDB() throws UnknownHostException {
-		MongoClient mongoClient = createMongoClient();
-		mongoTestHelper=new MongoTestHelper(mongoClient);
-		
-		mongoDatabase = new MongoDatabaseWrapper(mongoClient);
-	}
-
+public class MongoDatabaseWrapperTest extends AbstractMongoDatabaseWrapperTest {
+	@Override
 	public MongoClient createMongoClient() {
 		Fongo fongo = new Fongo("mongo server 1");
 		MongoClient mongoClient = fongo.getMongo();
 		return mongoClient;
-	}
-
-	@Test
-	public void testGetAllStudentsEmpty() {
-		assertTrue(mongoDatabase.getAllStudentsList().isEmpty());
-	}
-
-	@Test
-	public void testGetAllStudentsNotEmpty() {
-		mongoTestHelper.addStudent("1", "first");
-		mongoTestHelper.addStudent("2", "second");
-
-		assertEquals(2, mongoDatabase.getAllStudentsList().size());
-	}
-
-	@Test
-	public void testFindStudentByIdNotFound() {
-		mongoTestHelper.addStudent("1", "first");
-		assertNull(mongoDatabase.findStudentById("2"));
-	}
-
-	@Test
-	public void testFindStudentByIdFound() {
-		mongoTestHelper.addStudent("1", "first");
-		mongoTestHelper.addStudent("2", "second");
-
-		Student findStudentById = mongoDatabase.findStudentById("2");
-		assertNotNull(findStudentById);
-		assertEquals("2", findStudentById.getId());
-		assertEquals("second", findStudentById.getName());
 	}
 }
